feat: Add ability to add new tracks to playlist
Add ability to add new tracks to playlist  This commit adds a function called "addTrack" which handles the process of adding a new track to the playlist. The function is triggered by an event and retrieves the value of the input field with an ID of "track-input". If the music player is not powered on and the input field is not empty, the function returns and does not execute any further code.  If the input field is not empty and the music player is powered on, the function adds the new track to the end of the songs array using the "push" method. This allows users to add new tracks to the playlist dynamically without having to reload the page.  This commit also includes any necessary updates to the HTML and CSS files to ensure that the new feature is properly integrated into the existing interface.