feat: This code defines three functions and two event listener for an HTML select element:
drawExistingColors: Populates an HTML select element with options for each unique color value in an array of objects vehiclesInfo. drawExistingYears: Populates an HTML select element with options for each unique year value in vehiclesInfo. filterVehiclesByYear: Filters the vehiclesInfo array to only include objects with a specific year property value. The filtered array is then passed as an argument to the displayCard function. This function is triggered by the change event on the selectYearElement HTML element. An event listener is added to the selectYearElement element that listens for the change event and calls the filterVehiclesByYear function.